+++ aifd/cli/__init__.py
@@ -7,6 +7,7 @@ import click
 from aifd import __version__
 
 from .ai import ai
+from .cosmos import cosmos
 from .quota import quota
 from .vault import vault
 
@@ -22,5 +23,6 @@ def cli() -> None:
 
 
 cli.add_command(ai)
+cli.add_command(cosmos)
 cli.add_command(quota)
 cli.add_command(vault)